requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
psycopg2-binary
sqlparse
python-dotenv
rich
